+++ Source/MockUIAlertController/QCOMockAlertVerifier.h
@@ -22,6 +22,7 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nullable, nonatomic, copy) NSString *title;
 @property (nullable, nonatomic, copy) NSString *message;
 @property (nonatomic, assign) UIAlertControllerStyle preferredStyle;
+@property (nonatomic, copy) NSArray<UIAlertAction *> *actions;
 @property (nonatomic, readonly) NSArray *actionTitles;
 @property (nullable, nonatomic, strong) QCOMockPopoverPresentationController *popover;