+++ Source/MockUIAlertController/QCOMockAlertVerifier.h
@@ -24,7 +24,7 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic, assign) UIAlertControllerStyle preferredStyle;
 @property (nonatomic, copy) NSArray<UIAlertAction *> *actions;
 @property (nullable, nonatomic, strong) UIAlertAction *preferredAction;
-@property (nonatomic, readonly) NSArray *actionTitles;
+@property (nonatomic, readonly) NSArray *actionTitles DEPRECATED_MSG_ATTRIBUTE("Use actions");
 @property (nullable, nonatomic, strong) QCOMockPopoverPresentationController *popover;
 @property (nullable, nonatomic, copy) NSArray<UITextField *> *textFields;
 
@@ -39,7 +39,7 @@ NS_ASSUME_NONNULL_BEGIN
  * @abstract Returns the UIAlertActionStyle for the button with the specified title.
  * @discussion Throws an exception if no button with that title is found.
  */
-- (UIAlertActionStyle)styleForButtonWithTitle:(NSString *)title;
+- (UIAlertActionStyle)styleForButtonWithTitle:(NSString *)title DEPRECATED_MSG_ATTRIBUTE("Use actions");
 
 /*!
  * @abstract Executes the action for the button with the specified title.